feat: created base class BaseDbToolService to share common code with other DB Tool Services
